+++ tests/test_client.py
@@ -1,5 +1,6 @@
 import asyncio
 import copy
+import contextlib
 import io
 import json
 from collections import Counter
@@ -9,6 +10,7 @@ import asynctest
 import pytest
 
 from speechmatics import client
+from speechmatics.exceptions import ForceEndSession
 from speechmatics.models import (
     ConnectionSettings,
     ServerMessageType,
@@ -151,6 +153,68 @@ def test_middlewares_called(mock_server, mocker):
     assert all_handler.call_count == len(mock_server.messages_received)
 
 
+def test_force_end_session_from_event_handler(mock_server):
+    ws_client, transcription_config, audio_settings = default_ws_client_setup(
+        mock_server.url
+    )
+
+    def session_ender(event):
+        raise ForceEndSession
+
+    events = []
+    ws_client.add_event_handler("all", events.append)
+    ws_client.add_event_handler(ServerMessageType.RecognitionStarted,
+                                session_ender)
+
+    with open(path_to_test_resource("ch.wav"), "rb") as audio_stream:
+        ws_client.run_synchronously(
+            audio_stream, transcription_config, audio_settings)
+    mock_server.wait_for_clean_disconnects()
+
+    # Only one message should have been sent from the server
+    # (which is RecognitionStarted) before the session was
+    # forcefully ended.
+    assert len(mock_server.messages_sent) == 1
+    assert mock_server.messages_sent[0]["message"] == "RecognitionStarted"
+
+    # The client should only have recorded one event.
+    assert len(events) == 1
+
+
+@pytest.mark.parametrize(
+    "client_message_type, expect_received_count, expect_sent_count",
+    [
+        pytest.param(ClientMessageType.StartRecognition, 0, 1,
+                     id="StartRecognition"),
+        pytest.param(ClientMessageType.AddAudio, 1, 2,
+                     id="AddAudio"),
+    ]
+)
+def test_force_end_session_from_middleware(
+        mock_server, client_message_type,
+        expect_received_count, expect_sent_count):
+    ws_client, transcription_config, audio_settings = default_ws_client_setup(
+        mock_server.url
+    )
+
+    def session_ender(event, _):
+        raise ForceEndSession
+
+    sent_messages = []
+    ws_client.add_middleware("all",
+                             lambda msg, _: sent_messages.append(msg))
+    ws_client.add_middleware(client_message_type,
+                             session_ender)
+
+    with open(path_to_test_resource("ch.wav"), "rb") as audio_stream:
+        ws_client.run_synchronously(
+            audio_stream, transcription_config, audio_settings)
+    mock_server.wait_for_clean_disconnects()
+
+    assert len(mock_server.messages_received) == expect_received_count
+    assert len(sent_messages) == expect_sent_count
+
+
 def test_update_transcription_config_sends_set_recognition_config(mock_server):
     ws_client, transcription_config, audio_settings = default_ws_client_setup(
         mock_server.url
@@ -206,8 +270,13 @@ def test_helpful_error_message_received_on_connection_reset_error():
     """
     ws_client, _, _ = default_ws_client_setup("wss://this-url-wont-be-used:1")
 
-    async def mock_connect(*args, **kwargs):
+    @contextlib.asynccontextmanager
+    async def mock_connect(*_, **__):
         raise ConnectionResetError("foo")
+        # We need a yield here for this to be a valid context manager,
+        # even though the code is unreachable. Without it we get an error
+        # about a missing __anext__ attribute.
+        yield None  # pylint: disable=unreachable
 
     mock_logger_error_method = MagicMock()